move attestation cache to internal
